+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/properties/providers/metadata/SpringMetadataTypeReferenceProvider.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright © 2024 Explyt Ltd
+ *
+ * All rights reserved.
+ *
+ * This code and software are the property of Explyt Ltd and are protected by copyright and other intellectual property laws.
+ *
+ * You may use this code under the terms of the Explyt Source License Version 1.0 ("License"), if you accept its terms and conditions.
+ *
+ * By installing, downloading, accessing, using, or distributing this code, you agree to the terms and conditions of the License.
+ * If you do not agree to such terms and conditions, you must cease using this code and immediately delete all copies of it.
+ *
+ * You may obtain a copy of the License at: https://github.com/explyt/spring-plugin/blob/main/EXPLYT-SOURCE-LICENSE.md
+ *
+ * Unauthorized use of this code constitutes a violation of intellectual property rights and may result in legal action.
+ */
+
+package com.explyt.spring.core.properties.providers.metadata
+
+import com.intellij.psi.ElementManipulators
+import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiReference
+import com.intellij.psi.PsiReferenceProvider
+import com.intellij.psi.impl.source.resolve.reference.impl.providers.JavaClassReferenceProvider
+import com.intellij.psi.impl.source.resolve.reference.impl.providers.JavaClassReferenceSet
+import com.intellij.util.ProcessingContext
+
+class SpringMetadataTypeReferenceProvider : PsiReferenceProvider() {
+
+    private val referenceProvider = JavaClassReferenceProvider()
+
+    override fun getReferencesByElement(element: PsiElement, context: ProcessingContext): Array<PsiReference> {
+        val valueText = ElementManipulators.getValueText(element)
+        val offset = ElementManipulators.getOffsetInElement(element)
+
+        return JavaClassReferenceSet(valueText, element, offset, false, referenceProvider)
+            .references
+    }
+
+}